backend/requirements.txt
fastapi
uvicorn
pymongo
redis
sqlalchemy
psycopg2-binary
transformers
google-generativeai
elasticsearch
python-dotenv
